test(connectors): collapse random source liveness helper
