--- core/integration/tests/connectors/source_suite.rs
@@ -1,97 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-use std::time::Duration;
-
-use iggy_common::{Consumer, Identifier, IggyMessage, MessageClient, PollingStrategy};
-use integration::harness::{TestHarness, seeds};
-use tokio::time::{sleep, timeout};
-
-pub(crate) struct SourceSuiteConfig {
-    pub(crate) consumer_name: &'static str,
-    pub(crate) min_messages: usize,
-    pub(crate) poll_batch: u32,
-    pub(crate) retry_interval: Duration,
-    pub(crate) timeout: Duration,
-}
-
-impl Default for SourceSuiteConfig {
-    fn default() -> Self {
-        Self {
-            consumer_name: "source_suite_consumer",
-            min_messages: 1,
-            poll_batch: 100,
-            retry_interval: Duration::from_millis(100),
-            timeout: Duration::from_secs(5),
-        }
-    }
-}
-
-pub(crate) async fn poll_until_min_messages(
-    harness: &TestHarness,
-    config: &SourceSuiteConfig,
-) -> Vec<IggyMessage> {
-    let client = harness.root_client().await.expect("root client");
-    let stream_id: Identifier = seeds::names::STREAM.try_into().unwrap();
-    let topic_id: Identifier = seeds::names::TOPIC.try_into().unwrap();
-    let consumer_id: Identifier = config.consumer_name.try_into().unwrap();
-
-    let poll = async {
-        let mut messages = Vec::new();
-
-        loop {
-            if let Ok(mut polled) = client
-                .poll_messages(
-                    &stream_id,
-                    &topic_id,
-                    None,
-                    &Consumer::new(consumer_id.clone()),
-                    &PollingStrategy::next(),
-                    config.poll_batch,
-                    true,
-                )
-                .await
-            {
-                messages.append(&mut polled.messages);
-
-                if messages.len() >= config.min_messages {
-                    return messages;
-                }
-            }
-
-            sleep(config.retry_interval).await;
-        }
-    };
-
-    timeout(config.timeout, poll).await.unwrap_or_else(|_| {
-        panic!(
-            "source suite timed out after {:?} waiting for {} messages for consumer {}",
-            config.timeout, config.min_messages, config.consumer_name
-        )
-    })
-}
-
-pub(crate) async fn assert_source_produces_messages(harness: &TestHarness) -> Vec<IggyMessage> {
-    let messages = poll_until_min_messages(harness, &SourceSuiteConfig::default()).await;
-    assert!(
-        !messages.is_empty(),
-        "source suite expected at least one message"
-    );
-    messages
-}
